server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Gemini", email = "gemini@google.com"},
]
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "passlib",
    "bcrypt",
    "python-jose[cryptography]",
    "pydantic-settings"
]
requires-python = ">=3.9"